refactor: fixed naming and extracted to type file
