+++ yarn-project/acir-simulator/src/client/client_execution_context.ts
@@ -10,7 +10,7 @@ import {
   toAcvmL1ToL2MessageLoadOracleInputs,
 } from '../acvm/index.js';
 import { NoteLoadOracleInputs, DBOracle } from './db_oracle.js';
-import { PackedArgsCache } from '../packedArgsCache.js';
+import { PackedArgsCache } from '../packed_args_cache.js';
 
 /**
  * A type that wraps data with it's read request index
