+++ yarn-project/acir-simulator/src/packed_args_cache.ts
