+++ yarn-project/types/src/index.ts
@@ -18,3 +18,4 @@ export * from './logs/tx_l2_logs.js';
 export * from './tx_execution_request.js';
 export * from './logs/l2_logs_source.js';
 export * from './partial_contract_address.js';
+export * from './packed_arguments.js';
